End Fix issue #13
To fit our needs, a collision in memoryReference must result to an
overwrite of the ancient value. The insert method of std::map
doesn't do this, so we use the overloading of the [] operator
instead.

symVarMemoryReference becomes a map from a symVar ID to a memory
address. This way correspond to its future use.

I change the signature to quickly of isMemoryReference in the last
commit, it was an error.
The value -1 is used as a NIL value and so this method can
definitively not return an unsigned int!

+++ src/includes/SymbolicEngine.h
@@ -54,7 +54,7 @@ class SymbolicEngine {
     std::map<uint64_t, uint64_t> memoryReference;
 
     /*
-     * Addresses -> Z3 Symbolic Variable
+     * Z3 Symbolic Variable -> Addresses
      * item1: memory address
      * item2: symbolic variable ID
      */
@@ -73,7 +73,7 @@ class SymbolicEngine {
     uint64_t              symbolicReg[25];
 
     /* public methods */
-    uint64_t               isMemoryReference(uint64_t addr);
+    int32_t               isMemoryReference(uint64_t addr);
     std::string           getSmt2LibVarsDecl();
     symbolicElement       *getElementFromId(uint64_t id);
     symbolicElement       *newSymbolicElement(std::stringstream &src);
